platform-checks: Move RenameCluster/Replica to own file
to be consistent with the other rename_* checks

Also fix the minimum version requirement as v0.57.3 is out without my
new platform-checks
