+++ misc/python/materialize/checks/all_checks.py
@@ -45,7 +45,9 @@ from materialize.checks.peek_cancellation import *  # noqa: F401 F403
 from materialize.checks.pg_cdc import *  # noqa: F401 F403
 from materialize.checks.range import *  # noqa: F401 F403
 from materialize.checks.regex import *  # noqa: F401 F403
+from materialize.checks.rename_cluster import *  # noqa: F401 F403
 from materialize.checks.rename_index import *  # noqa: F401 F403
+from materialize.checks.rename_replica import *  # noqa: F401 F403
 from materialize.checks.rename_source import *  # noqa: F401 F403
 from materialize.checks.rename_table import *  # noqa: F401 F403
 from materialize.checks.rename_view import *  # noqa: F401 F403
