+++ misc/python/materialize/checks/replica.py
@@ -11,7 +11,6 @@ from typing import List
 
 from materialize.checks.actions import Testdrive
 from materialize.checks.checks import Check
-from materialize.util import MzVersion
 
 
 class CreateReplica(Check):
@@ -98,60 +97,6 @@ class DropReplica(Check):
                 6
 
                 > SELECT * FROM drop_replica_view;
-                6
-           """
-            )
-        )
-
-
-class RenameReplica(Check):
-    def _can_run(self) -> bool:
-        return self.base_version >= MzVersion.parse("0.57.0-dev")
-
-    def manipulate(self) -> List[Testdrive]:
-        return [
-            Testdrive(dedent(s))
-            for s in [
-                """
-                > CREATE TABLE rename_replica_table (f1 INTEGER);
-                > INSERT INTO rename_replica_table VALUES (1);
-
-                > CREATE CLUSTER rename_replica REPLICAS ();
-
-                > SET cluster=rename_replica
-                > CREATE DEFAULT INDEX ON rename_replica_table;
-                > CREATE MATERIALIZED VIEW rename_replica_view AS SELECT COUNT(f1) FROM rename_replica_table;
-
-                > INSERT INTO rename_replica_table VALUES (2);
-                > CREATE CLUSTER REPLICA rename_replica.replica1 SIZE '2-2';
-                > INSERT INTO rename_replica_table VALUES (3);
-                > CREATE CLUSTER REPLICA rename_replica.replica2 SIZE '2-2';
-                > INSERT INTO rename_replica_table VALUES (4);
-                > ALTER CLUSTER REPLICA rename_replica.replica1 RENAME TO replica_new1
-                """,
-                """
-                > INSERT INTO rename_replica_table VALUES (5);
-                > ALTER CLUSTER REPLICA rename_replica.replica2 RENAME TO replica_new2;
-                > INSERT INTO rename_replica_table VALUES (6);
-                """,
-            ]
-        ]
-
-    def validate(self) -> Testdrive:
-        return Testdrive(
-            dedent(
-                """
-                > SET cluster=rename_replica
-
-                > SELECT * FROM rename_replica_table;
-                1
-                2
-                3
-                4
-                5
-                6
-
-                > SELECT * FROM rename_replica_view;
                 6
            """
             )
